requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Audio processing
torch==2.1.1
torchaudio==2.1.1
librosa==0.10.1
soundfile==0.12.1

# NeMo and related dependencies (simplified)
nemo_toolkit==1.22.0
omegaconf==2.3.0
hydra-core==1.3.2

# CTC decoding (optional)
pyctcdecode==0.5.0
kenlm==0.1

# Numeric computing
numpy==1.24.3
scipy==1.11.4

# Utilities
typing-extensions==4.8.0
python-dotenv==1.0.0
Cython
# Optional: For better performance
numba==0.58.1